requirements.txt
flask
flask-sqlalchemy
sqlalchemy
flask-smorest
python-dotenv
marshmallow
gunicorn